GetWorldFPS関数の型をint→floatに変更

+++ Source/GhostBurster/Private/Enemy/Enemys.cpp
@@ -79,7 +79,7 @@ void AEnemys::EnemyDead()
 }
 
 //Œ»Ý‚ÌFPS‚ðŽæ“¾‚·‚é
-int AEnemys::GetWorldFPS()
+float AEnemys::GetWorldFPS()
 {
 	//DeltaTimeŽæ“¾
 	float DeltaTime = GetWorld()->GetDeltaSeconds();
